+++ lib/src/models/id_document_type.dart
@@ -1,1 +1,11 @@
-enum IdDocumentType { aadhaar, pan, drivingLicense, gstin, voterId, passport }
+enum IdDocumentType {
+  aadhaar,
+  pan,
+  drivingLicense,
+  gstin,
+  voterId,
+  passport,
+  pinCode,
+  email,
+  phone,
+}